Cleanup reflection header
