+++ source/library/reflect/public/potato/reflect/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(reflect PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/_export.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/_wrapper.h
     ${CMAKE_CURRENT_SOURCE_DIR}/reflect.h
     ${CMAKE_CURRENT_SOURCE_DIR}/stream_serializer.h
 )
